Update it to latest commit

+++ pages/login.php
@@ -1,14 +1,15 @@
 <?php
 //  AcmlmBoard XD - Login page
 //  Access: guests
+if (!defined('BLARG')) die();
 
 if($_POST['action'] == "logout")
 {
-	setcookie("logsession", "", 2147483647, BOARD_ROOT, "", false, true);
+	setcookie("logsession", "", 2147483647, URL_ROOT, "", false, true);
 	Query("UPDATE {users} SET loggedin = 0 WHERE id={0}", $loguserid);
 	Query("DELETE FROM {sessions} WHERE id={0}", doHash($_COOKIE['logsession'].SALT));
 
-	die(header("Location: ".BOARD_ROOT));
+	die(header("Location: ".URL_ROOT));
 }
 elseif(isset($_POST['actionlogin']))
 {
@@ -25,7 +26,7 @@ elseif(isset($_POST['actionlogin']))
 	
 	// auth plugins
 	if (!$okay)
-		{ $bucket = 'login'; include('lib/pluginloader.php'); }
+		{ $bucket = 'login'; include(BOARD_ROOT.'lib/pluginloader.php'); }
 
 	if(!$okay)
 	{
@@ -37,7 +38,7 @@ elseif(isset($_POST['actionlogin']))
 		//TODO: Tie sessions to IPs if user has enabled it (or probably not)
 
 		$sessionID = Shake();
-		setcookie("logsession", $sessionID, 2147483647, BOARD_ROOT, "", false, true);
+		setcookie("logsession", $sessionID, 2147483647, URL_ROOT, "", false, true);
 		Query("INSERT INTO {sessions} (id, user, autoexpire) VALUES ({0}, {1}, {2})", doHash($sessionID.SALT), $user['id'], $_POST['session']?1:0);
 
 		Report("[b]".$user['name']."[/] logged in.", 1);
@@ -58,7 +59,7 @@ elseif(isset($_POST['actionlogin']))
 		if (count($matches) > 0)
 			Query("INSERT INTO {passmatches} (date,ip,user,matches) VALUES (UNIX_TIMESTAMP(),{0},{1},{2})", $_SERVER['REMOTE_ADDR'], $user['id'], implode(',',$matches));
 
-		die(header("Location: ".BOARD_ROOT));
+		die(header("Location: ".URL_ROOT));
 	}
 }
